Refactor attestation collection
Instead of storing attestation collections as a map of Type to Attestor
I modified it to hold an array of objects with a type and attestor key.
There could potentially be circumstances where having multiple
attestations of the same type could happen, though I can't imagine why
right now.  The real reason I made the switch was because it maintains
the order in which the attestations were executed.