attestation/commandrun: handle signal injection on non-ptrace signals
https://man7.org/linux/man-pages/man2/ptrace.2.html specifically the
"Signal-delivery-stop" and "Signal injection and suppression" sections
are relevant here.  @colek42 noticed that witness froze while tracing
a build of the linux kernel.  I also encountered this and got caught up
in tracking down what was going on.

What I found is that the freeze always followed a child process calling
the `exit_group` syscall and a parent hanging on a `SIGCHLD` signal.
After a bit more of a thorough reading of the ptrace man page I found
that the signal option being passed to the `PtraceSyscall` function was
suppressing the SIGCHLD signals and causing the parent processes to
hang.

By setting `PTRACE_O_TRACESYSGOOD` we can differentiate between signals
triggered by our tracer and other signals.  This was already set but I
wasn't using that functionality before.  The code now inspects the
signal and if it is not a ptrace SIGTRAP we re-inject the signal back to
the process for it to handle properly.  The linux kernel now builds fine
with witness.