feat: evaluate rego policies during witness policy verification
Adds support for base64 encoded rego policies within a witness policy.
Rego policies may be added to any attestation in a step.  Currently rego
policies are expected to have a data value called 'deny' which is a
string array.  If any strings are in this array the policy will be
considered denied.  This allows policies to tell witness specifics about
why the policy failed.

The witness rego handling code will wrap all of the policies for a step
in a singular rego evaluation, and grab any deny reasons from the policy
to be returned to the caller.  An example policy may look something like
below:

```
package commandrun.cmd

deny[msg] {
    input.cmd != ["go", "build", "./"]
    msg := "unexpected cmd"
}
```

Signed-off-by: Mikhail Swift <mswift@mswift.dev>