feat: retrieve attestations from rekor
This is a fairly massive re-work of the verification code.

Tying our policy verification code to the DSSE package was something I
wasn't a fan of to begin with -- there's an existing issue to purge that
out and refactor the policy code.

So, this is the first step of that process. This breaks out all of the
DSSE parsing and verification from the policy package. The policy
package now expects all incoming in-toto statements to have had their
signatures verified prior to being called.

This relies on some (as of now) yet to be merged rekor functionality
(issue sigstore/rekor##582).

Envelopes are now verified against all of the policy's public keys and
trust bundles prior to policy.Verify being called.  policy.Verify will
still ensure the verified in-toto statement was signed by a trusted
functionary for each step.