feat!: add more robust policy constraints
Based off of in-toto ITE7 regarding PKI support for in-toto layouts. We
wrote the code for in-toto-golang, so there's a lot of crossover.

Adds support for constraining allowed x509 certificates by attributes on
the certificate itself. Example use cases may be constraining on an
expected SPIFFE ID or user's email to ensure that the identity for the
certificate is who we trust as a functionary.

BREAKING CHANGE: policies that used certificate constraints previously
may now fail due to the additional constraints in this pull request.
Since the unmarshal code will default these constraints to empty arrays,
any certificate will fail the constraint unless their respective
attributes are also empty.

Signed-off-by: Mikhail Swift <mikhail@testifysec.com>