fix: use intermediates on dsse signature to verify certificate
Please note that the Certificate and Intermediates fields on the DSSE
signature object are not spec standard fields and are custom to Witness
currently.

Currently the Witness run code is adding additional intermediates that
may be required to verify trust back to a root. This is useful when
using SPIFFE or other cases where an intermediate may be rotated fairly
frequently and cannot be embedded in the policy itself. However, the
witness verification code was not taking these additional intermediates
into account causing signatures to fail verification.

This change makes sure the DSSE code appropriately adds these
intermediates to the intermediate pool while verifying signatures with
x509 certificates.

Signed-off-by: Mikhail Swift <mikhail@testifysec.com>