chore: add logging when an unknown flag is provided
This is happening due to the order in which Cobra is running functions.
Previously we were initializing our logging library in a Cobra
OnInitialize function but this function gets run after Cobra has parsed
the command line flags.

This resulted in our logging not being appropriately initialized so the
error was being suppressed. By initializing our logger in the New
function and only setting the level in the OnInitialize the unknown flag
error gets logged appropriately.

Signed-off-by: Mikhail Swift <mikhail@testifysec.com>