feat: Enable Witness Policy verify from Archivista
Enables Witness to retrieve Policy from Archivista.

If a user provides a Policy `gitoid`, Witness will attempt to
retrieve it from Archivista.

To maintain backward compatibility, Witness first tries to load
the file locally. If the local file loading fails and an
Archivista is configured, it will retrieve the Policy from
Archivista.

Signed-off-by: Kairo Araujo <kairo.araujo@testifysec.com>